requirements.txt
ipdb
pytest
pytest-random-order
ruff
pillow
